requirements.txt
openai-whisper>=20231117
torch>=2.0.0
transformers>=4.30.0
gradio>=3.50.0
pysrt>=1.1.2
yt-dlp>=2023.3.4
librosa>=0.10.0
ffmpeg-python>=0.2.0
sentencepiece>=0.1.99
groq>=0.4.0
python-dotenv>=1.0.0
